refactor: naming
affects: @esri/hub-common
