+++ packages/common/src/urls/index.ts
@@ -13,7 +13,7 @@ export * from "./get-hub-url-from-portal";
 export * from "./get-item-home-url";
 export * from "./get-item-api-url";
 export * from "./get-item-data-url";
-export * from "./URLify";
+export * from "./convert-urls-to-anchor-tags";
 
 const MAP_OR_FEATURE_SERVER_URL_REGEX = /\/(map|feature)server/i;
