+++ packages/common/test/urls/convert-urls-to-anchor-tags.test.ts
@@ -1,9 +1,9 @@
-import { URLify } from "../../src";
+import { convertUrlsToAnchorTags } from "../../src";
 
-describe("URLify", function () {
+describe("convertUrlsToAnchorTags", function () {
   it("converts the url in a string to hyperlink", function () {
     const content = "Hello, https://foo.bar is supposed to be a hyperlink";
-    expect(URLify(content)).toBe(
+    expect(convertUrlsToAnchorTags(content)).toBe(
       'Hello, <a target="_blank" href="https://foo.bar">https://foo.bar</a> is supposed to be a hyperlink'
     );
   });
@@ -11,7 +11,7 @@ describe("URLify", function () {
   it("converts multiple urls in a string to hyperlinks", function () {
     const content =
       "Hello, https://foo.baz and https://bar.baz are supposed to be hyperlinks";
-    expect(URLify(content)).toBe(
+    expect(convertUrlsToAnchorTags(content)).toBe(
       'Hello, <a target="_blank" href="https://foo.baz">https://foo.baz</a> and <a target="_blank" href="https://bar.baz">https://bar.baz</a> are supposed to be hyperlinks'
     );
   });
